MAKING THE CODE MORE C++ COMPLIANT (USING math.h AND limits.h)

+++ Toolkit/Maths/Matrix2D.cpp
@@ -1,5 +1,6 @@
 #include "Matrix2D.h"
-#include <math.h>
+
+#include <cmath>
 
 using namespace std;
 using namespace ci;
@@ -27,8 +28,8 @@ void Matrix2D::scale(float s)
 
 void Matrix2D::rotate(float a)
 {
-	float c = cosf(a);
-	float s = sinf(a);
+	float c = cos(a);
+	float s = sin(a);
     
     float temp1 = m00;
     float temp2 = m01;
